ui: change the font size in charts to 12.5sp

+++ app/src/main/java/com/maxpoliakov/skillapp/util/ui/Fonts.kt
@@ -2,7 +2,7 @@ package com.maxpoliakov.skillapp.util.ui
 
 import android.content.Context
 
-fun Int.sp(context: Context): Float {
+fun Float.sp(context: Context): Float {
     val px = this * context.resources.displayMetrics.scaledDensity
     return (px / context.resources.displayMetrics.density)
 }
